+++ CHANGELOG.md
@@ -45,6 +45,7 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 - Add missing group chat channel message when a user is added to the group.
 - Add missing group chat channel message when a user is kicked from the group.
 - Add missing group chat channel message when a user is promoted in the group.
+- Handle TIMESTAMPTZ return types in Lua runtime custom SQL queries.
 
 ## [2.6.0] - 2019-07-01
 ### Added